+++ R/helper.R
@@ -6,11 +6,12 @@
 #' @template param_r_min
 #' @template param_r_max
 #' @template param_eta
+#' @template param_n_instances
 #' @template param_integer_budget
 #'
 #' @return [data.table::data.table()]
 #' @export
-hyperband_schedule = function(r_min, r_max, eta, integer_budget = FALSE) {
+hyperband_schedule = function(r_min, r_max, eta, n_instances = 1, integer_budget = FALSE) {
   r = r_max / r_min
   s_max = floor(log(r, eta))
   b = (s_max + 1) * r
@@ -19,7 +20,7 @@ hyperband_schedule = function(r_min, r_max, eta, integer_budget = FALSE) {
     nb = ceiling((b / r) * ((eta^s) / (s + 1)))
     rb = r * eta^(-s)
     map_dtr(0:s, function(i) {
-      ni = floor(nb * eta^(-i))
+      ni = floor(nb * eta^(-i)) * n_instances
       ri = r_min * rb * eta^i
       if (integer_budget) ri = round(ri)
       data.table(bracket = s, stage = i, budget = ri, n = ni)
@@ -35,15 +36,16 @@ hyperband_schedule = function(r_min, r_max, eta, integer_budget = FALSE) {
 #' @template param_r_min
 #' @template param_r_max
 #' @template param_eta
+#' @template param_n_instances
 #'
 #' @return `integer(1)`
 #' @export
-hyperband_n_configs = function(r_min, r_max, eta) {
+hyperband_n_configs = function(r_min, r_max, eta, n_instances = 1) {
   r = r_max / r_min
   s_max = floor(log(r, eta))
   budget = (s_max + 1) * r
 
-  sum(ceiling((budget / r) * (eta^(s_max:0)) / (s_max:0 + 1)))
+  sum(ceiling((budget / r) * (eta^(s_max:0)) / (s_max:0 + 1))) * n_instances
 }
 
 #' @title Hyperband Budget
@@ -54,11 +56,12 @@ hyperband_n_configs = function(r_min, r_max, eta) {
 #' @template param_r_min
 #' @template param_r_max
 #' @template param_eta
+#' @template param_n_instances
 #' @template param_integer_budget
-#' 
+#'
 #' @return `integer(1)`
 #' @export
-hyperband_budget = function(r_min, r_max, eta, integer_budget = FALSE) {
+hyperband_budget = function(r_min, r_max, eta, n_instances, integer_budget = FALSE) {
   schedule = hyperband_schedule(r_min, r_max, eta, integer_budget)
-  sum(schedule[, get("budget") * get("n")])
+  sum(schedule[, get("budget") * get("n")]) * n_instances
 }