+++ man-roxygen/param_n_instances.R
@@ -1,0 +1,2 @@
+#' @param n_instances (`integer(1)`)\cr
+#' Number of instances.